don't hate me :x

+++ ESp8266_socketIO/typedefs.h
@@ -20,5 +20,6 @@ typedef struct {
   char name[32];
   char type[32];
   bool on_state;
-  uint8_t brightness; // Int has 4 byte uin8_t only 1
+  int brightness;
+  // uint8_t brightness; // Int has 4 byte uin8_t only 1
 } DataPacket_t;
